- 타입 Person, CoffeeShop 개별파일로 생성
